+++ src/app/modules/index.ts
@@ -5,5 +5,6 @@ export * from './alarms/alarms.module'
 export * from './skresources/resources.module'
 export * from './map/map.module'
 export * from './utils/utils.module'
+export * from './experiments/experiments.module'
 
 export * from './skstream/skstream.module'