+++ src/app/modules/skresources/resource-dialogs.ts
@@ -1,13 +1,13 @@
 /** Resource Dialog Components **
 ********************************/
 
-import {Component, OnInit, Inject} from '@angular/core';
-import { SignalKClient } from 'signalk-client-angular';
+import { Component, OnInit, Inject, ChangeDetectorRef } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { MatBottomSheetRef, MAT_BOTTOM_SHEET_DATA } from '@angular/material/bottom-sheet';
 import { CdkDragDrop, moveItemInArray } from '@angular/cdk/drag-drop';
 import { Convert } from 'src/app/lib/convert'
 import { AppInfo } from 'src/app/app.info'
+import { SignalKClient } from 'signalk-client-angular';
 
 /********* ResourceDialog **********
 	data: {
@@ -502,8 +502,7 @@ export class ActiveResourcePropertiesModal implements OnInit {
         @Inject(MAT_BOTTOM_SHEET_DATA) public data: any) {
     }
     
-    ngOnInit() {
-        //console.log('data:', this.data);     
+    ngOnInit() {   
         if(this.data.resource[1].feature && this.data.resource[1].feature.geometry.coordinates) {
             if(this.data.type=='route'){
                 this.points= this.data.resource[1].feature.geometry.coordinates;
@@ -826,4 +825,170 @@ export class ChartLayersDialog implements OnInit {
     close() { this.dialogRef.close() } 
     
     isLocal(url:string) { return (url && url.indexOf('signalk')!=-1) ? 'map' : 'language' }
+}
+
+/********* TracksModal **********
+	data: {
+        title: "<string>" title text
+        skres: SKResource
+    }
+***********************************/
+@Component({
+	selector: 'ap-tracks-modal',
+	template: `
+        <div class="_ap-tracks">
+            <mat-toolbar>
+                <span>
+                    <button mat-icon-button 
+                        [disabled]="app.config.selections.tracks.length==0"
+                        matTooltip="Clear selections" matTooltipPosition="right"
+                        (click)="clearSelections()">
+                        <mat-icon>check_box_outline_blank</mat-icon>
+                    </button>
+                </span>
+                <span style="flex: 1 1 auto; padding-left:20px;text-align:center;">
+                    {{data.title}}
+                </span>
+                <span>
+                    <button mat-icon-button (click)="closeModal()"
+                        matTooltip="Close" matTooltipPosition="below">
+                        <mat-icon>keyboard_arrow_down</mat-icon>
+                    </button>
+                </span>
+            </mat-toolbar>          
+
+            <mat-card *ngFor="let trk of trackList; let idx= index">
+                <div style="display:flex;flex-wrap:no-wrap;">
+                    <div style="width:45px;">
+                        <mat-checkbox color="primary"
+                            [checked]="selTrk[idx]"
+                            (change)="handleCheck($event.checked, trk[0], idx)"
+                        ></mat-checkbox>
+                    </div>
+                    <div style="flex:1 1 auto;">
+                        <div class="key-label">
+                            {{trk[1].feature.properties.name}}
+                        </div>
+                        <div class="key-desc">
+                            {{trk[1].feature.properties.description}}
+                        </div>
+                    </div>
+                    <div style="width:45px;">
+                        <button mat-icon-button color="warn"
+                            (click)="handleDelete(trk[0], idx)">
+                            <mat-icon>delete</mat-icon>
+                        </button>
+                    </div>                    
+                </div>
+            </mat-card>
+        </div>	
+    `,
+    styles: [`  ._ap-tracks {
+                    font-family: arial;
+                    min-width: 300px;
+                }
+                ._ap-tracks .key-label {
+                    font-weight: 500;
+                }  
+                ._ap-tracks .key-desc {
+                    font-style: italic;
+                }                                	
+			`]
+})
+export class TracksModal implements OnInit {
+
+    public trackList:any;
+    public selTrk= [];
+
+    constructor(
+        public app: AppInfo,
+        private cdr: ChangeDetectorRef,
+        private sk: SignalKClient,        
+        public modalRef: MatBottomSheetRef<TracksModal>,
+        @Inject(MAT_BOTTOM_SHEET_DATA) public data: any) {
+    }
+    
+    ngOnInit() { 
+        if(this.data.title==='undefined') { this.data['title']= 'Tracks' }
+        this.getTracks();
+    } 
+
+    closeModal() { this.modalRef.dismiss() }    
+
+    getTracks() {
+        this.sk.api.get('resources/tracks').subscribe(
+            (trks:any)=> { 
+                this.trackList= Object.entries(trks).map( trk=> {
+                    trk[1]['feature']['id']= trk[0].toString();
+                    delete(trk[1]['$source']);
+                    delete(trk[1]['timestamp']);
+                    return trk;
+                });  
+                this.selTrk= [];             
+                for(let i=0;i<this.trackList.length;i++) {
+                    this.selTrk.push( 
+                        (this.app.config.selections.tracks.includes(this.trackList[i][0])) ?
+                        true : false
+                    );
+                }
+                this.cdr.detectChanges();             
+            },
+            ()=> { 
+                this.trackList= [];
+                this.selTrk= [];
+                this.cdr.detectChanges();
+            }
+        );        
+    }
+
+    handleDelete(id:string, idx:number) {
+        if(!this.data.skres) { return }
+        this.trackList= this.trackList.filter( t=> {
+            return (t[0]==id) ? false : true;
+        });         
+        this.data.skres.showTrackDelete(id).subscribe( ok=> { 
+            if(ok) {
+                let i= this.app.config.selections.tracks.indexOf(id);
+                if(i!=-1) { this.app.config.selections.tracks.splice(i,1) }   
+                this.data.skres.deleteTrack(id);
+                setTimeout(this.getTracks.bind(this), 2000);
+                this.app.saveConfig();
+            }
+            else { this.getTracks() }
+        });
+
+    }    
+
+    handleCheck(checked:boolean, id:string, idx:number) {
+        this.selTrk[idx]=checked;
+        if(checked) {
+            this.app.config.selections.tracks.push(id);
+        }
+        else {
+            let i= this.app.config.selections.tracks.indexOf(id);
+            if(i!=-1) { this.app.config.selections.tracks.splice(i,1) }
+        }
+        this.app.saveConfig();
+        this.updateTracks();
+    }
+
+    clearSelections() {
+        this.selTrk= [];
+        for(let i=0;i<this.trackList.length;i++) {
+            this.selTrk[i]= false;
+        }          
+        this.app.config.selections.tracks= [];
+        this.app.saveConfig();
+        this.updateTracks();
+    }
+    
+    updateTracks() {
+        this.app.data.tracks= this.trackList.map( trk=> {
+            return trk[1];
+        }).filter( t=> {
+            return (this.app.config.selections.tracks.includes(t.feature.id) ) ?
+                true : false;
+        });        
+        if(this.data.skres) { this.data.skres.trackSelected() }  
+    }
 }